Fix #15: Chronos2Backend.load() passes torch_dtype, not dtype
The backend was calling
    BaseChronosPipeline.from_pretrained(..., dtype=torch_dtype)
but the chronos pipeline's signature accepts torch_dtype= and rejects
dtype= (it forwards through to ChronosBoltModelForForecasting.__init__,
which has no dtype parameter).

Existing unit tests pass a mocked pipeline so they never see the wire-
level kwarg; the real failure only surfaced when running the
quickstart_batch.py example end-to-end against amazon/chronos-bolt-tiny.

Verified by re-running the example: 5 rows -> 5 forecasts in 39s with
plausible output.

Closes #15.

+++ src/sheaf/backends/chronos.py
@@ -78,7 +78,7 @@ class Chronos2Backend(ModelBackend):
         self._pipeline = BaseChronosPipeline.from_pretrained(
             self._model_id,
             device_map=self._device_map,
-            dtype=torch_dtype,
+            torch_dtype=torch_dtype,
         )
         self._is_bolt = isinstance(self._pipeline, ChronosBoltPipeline)
